+++ src/GMGPolar/writeToVTK.cpp
@@ -66,7 +66,7 @@ void GMGPolar::writeToVTK(const std::filesystem::path& file_path, const Level& l
     const PolarGrid& grid         = level.grid();
     const LevelCache& level_cache = level.levelCache();
 
-    assert(grid.numberOfNodes() == grid_function.size());
+    assert(static_cast<uint>(grid.numberOfNodes()) == grid_function.size());
 
     ConstVector<double> sin_theta_cache = level_cache.sin_theta();
     ConstVector<double> cos_theta_cache = level_cache.cos_theta();
